Cleaned code

+++ app/src/main/java/udacity/com/bakingtime/model/Ingredient.java
@@ -9,18 +9,18 @@ import android.os.Parcelable;
 
 public class Ingredient implements Parcelable{
 
-    private int mQuantity;
+    private double mQuantity;
     private String mMeasure;
     private String mName;
 
-    public Ingredient(int quantity, String measure, String name) {
+    public Ingredient(double quantity, String measure, String name) {
         this.mQuantity = quantity;
         this.mMeasure = measure;
         this.mName = name;
     }
 
     protected Ingredient(Parcel in) {
-        mQuantity = in.readInt();
+        mQuantity = in.readDouble();
         mMeasure = in.readString();
         mName = in.readString();
     }
@@ -37,11 +37,11 @@ public class Ingredient implements Parcelable{
         }
     };
 
-    public int getQuantity() {
+    public double getQuantity() {
         return mQuantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.mQuantity = quantity;
     }
 
@@ -68,7 +68,7 @@ public class Ingredient implements Parcelable{
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeInt(mQuantity);
+        dest.writeDouble(mQuantity);
         dest.writeString(mMeasure);
         dest.writeString(mName);
     }
